+++ runpod_sd_proxy/__init__.py
@@ -27,4 +27,4 @@ test = cur.execute("INSERT INTO model (use_model) VALUES ('v1-5-pruned-emaonly')
 db.commit()
 use_model = cur.lastrowid
 
-from .app import app
+from runpod_sd_proxy import routes
