+++ tests/test_app.py
@@ -1,6 +1,6 @@
 import unittest
 from unittest.mock import patch
-from runpod_sd_proxy.app import (
+from runpod_sd_proxy.routes import (
     pruned_sd_request,
     sdxl_sd_request,
     generate_image_based_on_model,
@@ -9,7 +9,7 @@ from runpod_sd_proxy.app import (
 
 class TestPrunedSdRequest(unittest.TestCase):
 
-    @patch("runpod_sd_proxy.app.requests.post")
+    @patch("runpod_sd_proxy.routes.requests.post")
     def test_pruned_sd_request(self, mock_post):
         # Mock response data
         mock_response_data = {
@@ -25,7 +25,7 @@ class TestPrunedSdRequest(unittest.TestCase):
         expected_image = "some_base64_encoded_data"
         self.assertEqual(result_image, expected_image)
 
-    @patch("runpod_sd_proxy.app.requests.post")
+    @patch("runpod_sd_proxy.routes.requests.post")
     def test_pruned_sd_request_error_on_invalid_response(self, mock_post):
         # Mock response data
         mock_response_data = {
@@ -44,7 +44,7 @@ class TestPrunedSdRequest(unittest.TestCase):
 
 class TestSDXLSdRequest(unittest.TestCase):
 
-    @patch("runpod_sd_proxy.app.requests.post")
+    @patch("runpod_sd_proxy.routes.requests.post")
     def test_sdxl_sd_request(self, mock_post):
         # Mock response data
         mock_response_data = {
@@ -60,7 +60,7 @@ class TestSDXLSdRequest(unittest.TestCase):
         expected_image = "some_base64_encoded_data"
         self.assertEqual(result_image, expected_image)
 
-    @patch("runpod_sd_proxy.app.requests.post")
+    @patch("runpod_sd_proxy.routes.requests.post")
     def test_sdxl_sd_request_error_on_invalid_response(self, mock_post):
         # Mock response data
         mock_response_data = {
@@ -79,7 +79,7 @@ class TestSDXLSdRequest(unittest.TestCase):
 
 class TestGenerateImageBasedOnModel(unittest.TestCase):
 
-    @patch("runpod_sd_proxy.app.requests.post")
+    @patch("runpod_sd_proxy.routes.requests.post")
     def test_generate_image_based_on_model_sd(self, mock_post):
         mock_response_data = {
             "output": {"images": ["some_base64_encoded_data"]},
@@ -94,7 +94,7 @@ class TestGenerateImageBasedOnModel(unittest.TestCase):
 
         self.assertEqual(output, expected_output)
 
-    @patch("runpod_sd_proxy.app.requests.post")
+    @patch("runpod_sd_proxy.routes.requests.post")
     def test_generate_image_based_on_model_sdxl(self, mock_post):
         mock_response_data = {
             "output": {"image_url": "data:image/png;base64,some_base64_encoded_data"},
